Add files via upload
AI call analyser for fraud calls.
This analysis the fraud calls and with help of subtitles helps person to understand the fraud. Basically in the new generation mobile phones we have audio reader to read the audio and caption it. this software uses that system to read the audio and help people understand the fraud .  It creates knowledge about banking and reduces the fraud chances.